create RoomData component and change box shadow scss
