+++ frontend/src/assets/styles/Room/create_room.scss
@@ -10,7 +10,7 @@
     }
 
     .card {
-        border-radius: 15px;
-        box-shadow: 2px 2px gray, 6px 6px black, 8px 8px dodgerblue;
+        border-radius: 10px;
+        box-shadow: rgba(0, 0, 0, 0.08) 0px 4px 12px;
     }
 }
