+++ frontend/src/assets/styles/Room/room_list.scss
@@ -11,12 +11,10 @@
     }
 
     .card {
-        border-radius: 15px;
-        display: inline-block;
+        border-radius: 10px;
         height: 150px;
-        margin: 1rem;
-        position: relative;
-        width: 260px;
-        box-shadow: 4px 4px gray, 6px 6px olive, 8px 8px dodgerblue;
+        margin: 20px;
+        width: 265px;
+        box-shadow: rgba(0, 0, 0, 0.08) 0px 4px 12px;
     }
 }
